Sleep before crashing when the HIDL HAL dies
The HAL crashing is an unrecoverable failure. Sometimes, the system
kills the HAL as part of a system shutdown, in which case Bluetooth
will also be killed. Sleeping for one second after being notified that
the HAL died is a way to make sure we don't report deaths that are a
result of the system killing the HAL out from under the Bluetooth
stack.

Bug: 294022818
Test: mma -j32
Test: Kill the hal, then Bluetooth with different timings
Google: 2740875
Change-Id: I9c17f4637639852965fdb529687a43ab26fde776